Add files via upload
CALL CENTER ANALYSIS
Background Of The Analysis:
A call center agent is a professional who handles customer interactions (calls, chats, emails, or messages) on behalf of a company. Their main role is to provide customer support, information, sales, or technical assistance.
The Dataset:
The dataset consists of 13 columns (id, customer_name, sentiment, csat_score, call_timestamp, reason, city, state, channel, response_time, call_duration_in_minutes, call_center) and 32941 rows.
Agents Overview:
In this dataset, the call center had 4 call centers and channels through which customers can reach out.
Customers Overview
There are 32941 Customers who interacted with the call center in the month of October, with most of them residing in the state of California.

Insight from the Analysis
1.	Customer Experience (sentiment, CSAT)
From the Analysis, 11,036 of the customers representing 33.58% of the total population, and 6,026 of the customers representing 18.29% of the total population were not pleased with the service of the call center, with NEGATIVE and Very Negative Sentiment leading the customers’ sentiment.
About 8,754 (26.57%) of the customers were not pleased nor displeased with the call center service. The Neutral Sentiment gave the most Customers Satisfaction (CSAT) Score, followed by the Negative sentiment.
2.	Operational metrics (response time, duration, SLA)
Out of the four Call-Centers, only Los Angeles/CA meet the response time (i.e response time fell below SLA time or Within SLA time) with only 11,995 customers experiencing this.
Customers prefer Call Center to other channels, Customers spent more time talking/reaching out  to agent through Call-Centers with total duration of 266,466.
3.	Volume and Demand Trends (time, geography, channel)
Customers get across to Call-Center agents mostly on Fridays, with total of 11,841 customers reaching out on Fridays for the month of October.
California is the location where most of the customers reached out from and Call-center is the most popular channel.

Recommendation:
From the analysis, most of the customers are not pleased with the service of the Call center, the analysis of the customers sentiment revealed that  about 7,000 customers were pleased with the service of the call center, and over 15,000 customers gave Negative and Very Negative sentiments. In this light, to improve the services of the call-center, this recommendation should be worked on:
Increase the number of call centers from four to about 10, and also increase the population of the agents, this will give room for timely intervention.